consulting_deck_generator/requirements.txt
python-pptx>=0.6.21
matplotlib>=3.6.0
seaborn>=0.11.0
google-generativeai>=0.8.0
pillow>=8.0.0